requirements.txt
arrow==1.2.3
asgiref==3.5.2
binaryornot==0.4.4
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
cookiecutter==2.1.1
coverage==7.2.7
Django==4.1.3
django-browser-reload==1.6.0
django-cors-headers==3.13.0
django-tailwind==3.4.0
exceptiongroup==1.1.2
factory-boy==3.3.0
Faker==19.2.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.1
packaging==23.1
Pillow==9.3.0
pluggy==1.2.0
pytest==7.4.0
pytest-cov==4.1.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python-slugify==7.0.0
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.7.1
urllib3==1.26.12
